# spec/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Example API v2",
    "description": "This document describes HTTP REST JSON API located at https://{{host}}/api/v2\nand used by **Example** mobile and desktop clients.\n\nAn actual version of this document is published online: https://docs.example.com/api/{{environment}}",
    "version": "{{version}}"
  },
  "host": "{{host}}",
  "basePath": "/api/v2",
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "securityDefinitions": {
    "token": {
      "name": "X-Access-Token",
      "type": "apiKey",
      "in": "header"
    }
  },
  "paths": {
...
    "/companies": {
      "post": {
        "summary": "**Create** new `Company`",
        "description": "**Create** new [`Company`](#/definitions/Company)",
        "operationId": "POST--companies",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean"
                },
                "id": {
                  "type": "integer"
                }
              },
              "required": [
                "success",
                "id"
              ]
            }
          }
        },
        "tags": [
          "Agent endpoints"
        ],
        "security": [
          {
            "token": []
          }
        ],
        "parameters": [
          {
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "company": {
                  "$ref": "#/definitions/CompanyNew"
                }
              },
              "required": [
                "company"
              ]
            },
            "in": "body"
          }
        ]
      }
    },
    "/companies/:id": {
      "get": {
        "summary": "**Retrieve** particular `Company`",
        "description": "**Retrieve** particular [`Company`](#/definitions/Company)",
        "operationId": "GET--companies--id",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "type": "object",
              "properties": {
                "company": {
                  "$ref": "#/definitions/Company"
                }
              },
              "required": [
                "company"
              ]
            }
          }
        },
        "tags": [
          "Agent endpoints"
        ],
        "security": [
          {
            "token": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "required": true,
            "type": "string",
            "in": "path"
          }
        ]
      },
      "patch": {
        "summary": "**Update** particular `Company`",
        "description": "**Update** particular [`Company`](#/definitions/Company)",
        "operationId": "PATCH--companies--id",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean"
                }
              },
              "required": [
                "success"
              ]
            }
          }
        },
        "tags": [
          "Agent endpoints"
        ],
        "security": [
          {
            "token": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "required": true,
            "type": "string",
            "in": "path"
          },
          {
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "company": {
                  "$ref": "#/definitions/CompanyUpdate"
                }
              },
              "required": [
                "company"
              ]
            },
            "in": "body"
          }
        ]
      },
      "delete": {
        "summary": "**Delete** particular `Company`",
        "description": "**Delete** particular [`Company`](#/definitions/Company)",
        "operationId": "DELETE--companies--id",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean"
                }
              },
              "required": [
                "success"
              ]
            }
          }
        },
        "tags": [
          "Agent endpoints"
        ],
        "security": [
          {
            "token": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "required": true,
            "type": "string",
            "in": "path"
          }
        ]
      }
    },
...
  },
...
  "definitions": {
    "Company": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "companyName": {
          "type": "string"
        },
        "companyStatus": {
          "type": [
            "string",
            "null"
          ]
        },
        "dormant": {
          "type": "boolean"
        },
        "industry": {
          "type": "string"
        },
        "size": {
          "type": "integer"
        },
        "status": {
          "type": "string"
        },
        "color": {
          "type": "string"
        },
        "logo": {
          "type": [
            "string",
            "null"
          ]
        },
        "dealStatus": {
          "type": "string"
        },
        "pipelineStage": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        }
      },
      "required": [
        "id",
        "companyName",
        "dormant",
        "industry",
        "size",
        "status",
        "color",
        "dealStatus",
        "pipelineStage",
        "createdAt",
        "updatedAt"
      ]
    },
    "CompanyNew": {
      "type": "object",
      "properties": {
        "companyName": {
          "type": "string"
        },
        "companyStatus": {
          "type": [
            "string",
            "null"
          ]
        },
        "dormant": {
          "type": "boolean"
        },
        "industry": {
          "type": "string"
        },
        "size": {
          "type": "integer"
        },
        "status": {
          "type": "string"
        },
        "color": {
          "type": "string"
        },
        "logo": {
          "type": [
            "string",
            "null"
          ]
        },
        "dealStatus": {
          "type": "string"
        },
        "pipelineStage": {
          "type": "string"
        }
      },
      "required": [
        "companyName",
        "dormant",
        "industry",
        "size",
        "status",
        "color",
        "dealStatus",
        "pipelineStage"
      ],
      "additionalProperties": false
    },
    "CompanyUpdate": {
      "type": "object",
      "properties": {
        "companyName": {
          "type": [
            "string",
            "null"
          ]
        },
        "companyStatus": {
          "type": [
            "string",
            "null"
          ]
        },
        "dormant": {
          "type": [
            "boolean",
            "null"
          ]
        },
        "industry": {
          "type": [
            "string",
            "null"
          ]
        },
        "size": {
          "type": [
            "integer",
            "null"
          ]
        },
        "status": {
          "type": [
            "string",
            "null"
          ]
        },
        "color": {
          "type": [
            "string",
            "null"
          ]
        },
        "logo": {
          "type": [
            "string",
            "null"
          ]
        },
        "dealStatus": {
          "type": [
            "string",
            "null"
          ]
        },
        "pipelineStage": {
          "type": [
            "string",
            "null"
          ]
        }
      },
      "additionalProperties": false
    },
...
  }
}